requirements.txt
requests=2.23.0
pysocks=1.7.1
stemp=1.8.0
selenium=3.141.0